Added an option to keep compiled binary by passing "-k keep"

+++ build.sh
@@ -9,10 +9,11 @@
 (Works on any Linux distro with gcc and chmod)
 '
 
-while getopts f: flag
+while getopts f:k: flag
 do
     case "${flag}" in
         f) file=${OPTARG};;
+        k) keep=${OPTARG};;
     esac
 done
 
@@ -26,6 +27,13 @@ gcc ${file} -Wall -Wextra -pedantic -o ${file}.tmp
 chmod +x ${file}.tmp
 } &> /dev/null
 ./${file}.tmp
-{
+
+if [[ "$keep" == "keep" ]]; then
+    true
+fi
+
+if [[ -z "$keep" ]]; then
+    {
     rm ${file}.tmp
-} &> /dev/null
+    } &> /dev/null
+fi
